+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 # @project-version@ (@build-time@)
 
 * Fixed a bug where the healthbar border did not reappear after being disabled and re-enabled, without needing a UI reload.
+* Fixed a bug where automatic spec-based profile switching did not work after login/reload until the options window had been opened once [GH-685].